Moved CSV to examples/CSV
